backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
PyJWT==2.1.0
Werkzeug==2.0.1
argon2-cffi==21.3.0
tabulate==0.9.0
Flask-Mail==0.9.1
gunicorn==20.1.0
certifi==2023.7.22
requests==2.28.2
cryptography==39.0.2
pycparser==2.21
SQLAlchemy==2.0.19
psutil==5.9.5
urllib3==1.26.16
Jinja2==3.0.3
itsdangerous==2.0.1
cffi==1.15.1
MarkupSafe==2.1.1
idna==3.4
six==1.16.0

# pip install -r requirements.txt